Several changes to speed up the code.
Now location_from_sentence.py doesn't have to load StanfordCoreNLP
every time it is run, but we need to run it as a server.

Reimplemented the models and indexed several columns, also added
code to calculate probabilities and added tables to store them,
but they are not being stored yet.